requirements.txt
asgiref==3.7.2
crispy-bootstrap5==0.7
dj-database-url==1.3.0
dj-email-url==1.0.6
Django==4.2.2
django-cache-url==3.4.4
django-crispy-forms==2.1
django-extensions==3.2.3
django-recaptcha==3.0.0
Django-Verify-Email==2.0.3
environs==9.5.0
gunicorn==20.1.0
marshmallow==3.20.1
packaging==23.2
psycopg2-binary==2.9.6
python-dotenv==1.0.0
sqlparse==0.4.4
typing_extensions==4.8.0
whitenoise==6.4.0
